==> 0 - train_configuration.ipynb <==
# Train Configuration Files

Cada treinamento vai ter um identificador hash que vai permear todo o processo.

import os
import pandas as pd
import pickle
import numpy as np

from AuxiliarFunctions import *

# Configurações gerais
hash_id = hash('treinamento_0')
label = 'Toy Data Classification with 10 StratifiedKFolds' # etiqueta do treinamento
raw_files_path = 'data/raw' # caminho para os arquivos raw (sem processamento)
processed_file_folder = 'data' # caminho para o arquivo único com todos os dados crus - processed_data.csv
processed_file_path =  os.path.join(processed_file_folder,'%s_processed_data.csv'%(hash_id)) # caminho para o arquivo único com todos os dados crus - processed_data.csv

# Configurações da Validação Cruzada
cv_alg = 'StratifiedKFolds' # StratifiedKFolds or LeaveOneOut
cv_folds = 10
cv_path = os.path.join(processed_file_folder,'indexes')

# Configurações de Pré-Processamento
preproc_alg = 'Não implementado para a aplicação!!!'

# Configurações de Pipeline
pipeline_path = os.path.join(processed_file_folder,'pipelines')
scaler_alg = 'StandardScaler'

# Arquivos de Treinamento
train_data_path = os.path.join(processed_file_folder,'%s_train_data.csv'%(hash_id)) # caminho para o arquivo único com todos os dados para treinamento - train_data_file.csv
train_trgt_path = os.path.join(processed_file_folder,'%s_trgt_data.csv'%(hash_id)) # caminho para o arquivo único com todos os alvos para treinamento - train_trgt_file.csv
model_path = os.path.join(processed_file_folder,'models') # caminho para onde os modelos serão enviados


dict_buffer = { 
    'hash_id':[hash_id],
    'label':[label],
    'raw_files_path':[raw_files_path],
    'processed_file_folder':[processed_file_folder],
    'processed_file_path':[processed_file_path],
    'cv_alg':[cv_alg],
    'cv_folds':cv_folds,
    'cv_path':cv_path,
    'preproc_alg':[preproc_alg],
    'pipeline_path':[pipeline_path],
    'scaler_alg':[scaler_alg],
    'train_data_path':[train_data_path],
    'train_trgt_path':[train_trgt_path],
    'model_path':[model_path],
}
df_buffer = pd.DataFrame(data=dict_buffer)

df_buffer.head()

config_file_path = 'data/config.csv'
if os.path.exists(config_file_path) == False:
    # arquivo não existe
    df_buffer.to_csv(config_file_path, index=False)
else:
    # arquivo existe
    df_old = pd.read_csv(config_file_path)
    if (hash_id in list(df_old['hash_id'])):
        # o hash_id já está salvo
        print('o hash_id já está salvo')
    else:
        df_old = pd.concat([df_old,df_buffer],axis=0, ignore_index=True)
        df_old.to_csv(config_file_path, index=False)

train_id = 0
df_config = pd.read_csv(config_file_path)

df_config.head(10)
