==> 1 - toydata_read_data.ipynb <==
# Toy Data Read Data    
Este notebook faz carga de dados do Toy.  
Aqui, a ideia é fazer um csv com as informações dos arquivos para ser lido sequencialmente


## Carga de bibliotecas

import os
import warnings
from scipy import signal
from scipy.signal import decimate, convolve, spectrogram, lfilter, cheby2, butter, cheb2ord, hilbert
from librosa import stft, fft_frequencies, frames_to_time, feature
import numpy as np
import math
import pandas as pd
import soundfile as sf

## Ler arquivo de configuração de treinamento

config_file_path = 'data/config.csv'
df_config = pd.read_csv(config_file_path)

train_id = 0

df_config

## Gerar os dados para treinamento

from sklearn.datasets import make_classification
rng = np.random.RandomState(0)

X, y = make_classification(n_samples=10000, 
                           n_features=20, 
                           n_informative=10, 
                           n_redundant=2, 
                           n_repeated=0, 
                           n_classes=2,#4, 
                           n_clusters_per_class=2, 
                           class_sep=1.0,
                           hypercube=True,
                           random_state=rng,
                           shuffle=True
                          )

df_buffer = pd.DataFrame(data=X)
df_buffer = df_buffer.add_prefix('feature_')
df_data = df_buffer.copy(deep=True)

df_target = pd.DataFrame(data=y, columns=['target'])
df_target.head()

#df_train = pd.concat([df_data, df_target], axis=1)
#df_train.to_csv(df_config['train_data_path'][train_id],index=False)

df_data.to_csv(df_config['train_data_path'][train_id],index=False)

df_target.to_csv(df_config['train_trgt_path'][train_id],index=False)
