==> 2 - cross_validation_generation.ipynb <==
# Cross Validation Index Generation

import pandas as pd
import numpy as np
import math
from sklearn.pipeline import Pipeline
from sklearn.model_selection import LeaveOneOut, StratifiedKFold
import pickle
import joblib
import os

## Load data

df_config = pd.read_csv('data/config.csv')
train_id = 0
df_config

### Load train data

df_data = pd.read_csv(df_config['train_data_path'][train_id])
df_trgt = pd.read_csv(df_config['train_trgt_path'][train_id])

df_data

# Stratified K-Folds

from sklearn.model_selection import StratifiedKFold
import pickle
import os

file_path = df_config['cv_path'][train_id]
n_folds = df_config['cv_folds'][train_id]

cv = StratifiedKFold(n_splits=n_folds)

#data = df_train.drop(columns=['target']).values
#target = df_train['target'].values

data = df_data
target = df_trgt

for ifold, (trn_idx, val_idx) in enumerate(cv.split(data,target)):
    cv_name ='%s_%s_CV_fold_%i_of_%i_cv_indexes.pkl'%(df_config['hash_id'][train_id],
                                                       df_config['cv_alg'][train_id],
                                                       ifold, n_folds)
    with open(os.path.join(file_path,cv_name),'wb') as file_handler:
        pickle.dump([trn_idx,val_idx],file_handler)

data.shape

target.shape
