==> 4 - train_models.ipynb <==
# Load Libraries

import os
import numpy as np
import pandas as pd
import pickle
import joblib

import Models
from AuxiliarFunctions import *

# Load Data

config_file_path = 'data/config.csv'
df_config = pd.read_csv(config_file_path)
train_id = 0

df_config

df_data = pd.read_csv(df_config['train_data_path'][train_id])
df_trgt = pd.read_csv(df_config['train_trgt_path'][train_id])

df_trgt

# Processo de Treinamento de um modelo simples

from sklearn.preprocessing import StandardScaler, MinMaxScaler
import time

import copy

model_type = 'SVM'
n_folds =  df_config['cv_folds'][train_id]

cv_path = df_config['cv_path'][train_id]
pipe_path = df_config['pipeline_path'][train_id]
model_path = df_config['model_path'][train_id]

print('Processing SVM Training')

kernels = ['linear', 'rbf']
regularizations = [0.5, 0.2, 0.9]


if os.path.exists(os.path.join(model_path, 'train_record.csv')):
    train_record = pd.read_csv(os.path.join(model_path, 'train_record.csv'))
else:
    train_record = None


for kernel in kernels:
    print('Processing Training for %s kernel'%(kernel))
    for regularization in regularizations:
        print('Processing Training for %s regularization'%(regularization))
        for ifold in range(n_folds):
            #if ifold >= 2: # para desenvolvimento
            #    break
            print('Training %i fold of %i folds\n\n'%(ifold+1, n_folds))

            print('Reading Cross-validation indexes')

            cv_name = '%s_%s_CV_fold_%i_of_%i_cv_indexes.pkl'%(df_config['hash_id'][train_id],
                                                               df_config['cv_alg'][train_id],
                                                               ifold, n_folds)
            print('Done')
            print('Reading Pipeline Object')
            with open(os.path.join(cv_path,cv_name),'rb') as file_handler:
                [trn_idx,val_idx] = pickle.load(file_handler)

            pipe_name ='%s_%s_CV_fold_%i_of_%i_pipe.pkl'%(df_config['hash_id'][train_id],
                                                               df_config['cv_alg'][train_id],
                                                               ifold, n_folds)
            with open(os.path.join(pipe_path,pipe_name),'rb') as file_handler:
                pipe = joblib.load(file_handler)

            print('Done')

            trn_data = pipe.transform(df_data)
            trn_trgt = df_trgt.values # tf.keras.utils.to_categorical(dev_target, num_classes=len(np.unique(dev_target)))

            print('Training for model')
            model_name = '%s_%s_%i_fold_model_%s_kernel_%s_regularization.pkl'%(df_config['hash_id'][train_id],
                                                                                model_type, ifold, kernel, 
                                                                                str(regularization).replace('.','-'))

            if os.path.exists(os.path.join(model_path, model_name)):
                print('Model is in %s'%(os.path.join(model_path, model_name)))
                
            else:
                print('No Model \n\n')

                model = Models.SVMClassificationModel(kernel=kernel, 
                                                      regularization=regularization, 
                                                      verbose=False)
                start_time = time.time()
                model.fit(trn_data, trn_trgt, trn_id=trn_idx, val_id=val_idx, random_state=0,)
                end_time = time.time() # in seconds

                predictions = model.predict(trn_data)
                df_predict = pd.DataFrame(data=np.concatenate((trn_trgt, 
                                                               predictions[:,np.newaxis]),
                                                              axis=1), 
                                          columns=['target', 'model_output'])
                prediction_name = copy.copy(model_name)
                prediction_name = prediction_name.replace('.pkl','_prediction_file.csv')
                df_predict.to_csv(os.path.join(model_path, prediction_name),index=False)
               
                model.save(os.path.join(model_path, model_name))
                
                acc = Models.acc_score(df_predict.loc[val_idx,'target'],
                                       df_predict.loc[val_idx,'model_output'])
                sens = Models.sensitivity_score(df_predict.loc[val_idx,'target'],
                                                df_predict.loc[val_idx,'model_output'])
                spec = Models.specificity_score(df_predict.loc[val_idx,'target'],
                                                df_predict.loc[val_idx,'model_output'])
                sp = Models.sp_index(df_predict.loc[val_idx,'target'],
                                     df_predict.loc[val_idx,'model_output'])
                auc = Models.auc_score(df_predict.loc[val_idx,'target'],
                                       df_predict.loc[val_idx,'model_output'])
                
                dict_train_record = {
                    'hash_id':[df_config['hash_id'][train_id]],'fold':[ifold],
                    'prediction_file':[prediction_name], 'kernel':[kernel],
                    'regularization':[regularization], 'Acc':[acc],
                    'Sens':[sens],'Spec':[spec],'SP':[sp], 'AUC':[auc],
                    'Time':[end_time-start_time]
                }
                if train_record is None:
                    train_record = pd.DataFrame(data=dict_train_record)
                else:
                    train_record = pd.concat([train_record,pd.DataFrame(data=dict_train_record)],axis=0, ignore_index=True)
train_record.to_csv(os.path.join(model_path,'%s_%s_train_record.csv'%(df_config['hash_id'][train_id], model_type)),index=False)

train_record = pd.read_csv(os.path.join(model_path,'%s_%s_train_record.csv'%(df_config['hash_id'][train_id], model_type)))
train_record

mean_grouped = train_record[['kernel','regularization',
                        'Acc','Sens','Spec','SP', 'AUC', 'Time']].groupby(['kernel', 'regularization']).mean()
std_grouped = train_record[['kernel','regularization',
                        'Acc','Sens','Spec','SP', 'AUC', 'Time']].groupby(['kernel', 'regularization']).std()
grouped = pd.concat([mean_grouped,std_grouped],axis=0)

grouped

choose_hyperparameters = {'criteria':['SP'],'kernel':['rbf'],'regularization':['0.9']}
df_choose_hyperparameters = pd.DataFrame(data=choose_hyperparameters)
df_choose_hyperparameters

df_choose_hyperparameters.to_csv(os.path.join(model_path,'%s_%s_choose_hyperparameters.csv'%(df_config['hash_id'][train_id], model_type)),index=False)
